requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.1.1
curated-tokenizers=0.0.9
curated-transformers=0.1.1
cymem=2.0.11
cython-blis=1.0.1
dataclasses=0.8
datasets=3.4.1
debugpy=1.8.13
decorator=5.2.1
huggingface_hub=0.29.3
hyperframe=6.1.0
idna=3.10
importlib-metadata=8.6.1
importlib_metadata=8.6.1
importlib_resources=6.5.2
iprogress=0.4
ipykernel=6.29.5
ipython=9.0.2
ipython_pygments_lexers=1.1.1
joblib=1.4.2
jupyter_client=8.6.3
jupyter_core=5.7.2
krb5=1.21.3
langcodes=3.3.0
libtorch=2.6.0
libutf8proc=2.10.0
libuv=1.50.0
libxml2=2.13.6
libzlib=1.3.1
llvm-openmp=19.1.7
lz4-c=1.10.0
markdown-it-py=3.0.0
markupsafe=3.0.2
matplotlib-inline=0.1.7
mdurl=0.1.2
mpc=1.3.1
mpfr=4.2.1
mpmath=1.3.0
multidict=6.2.0
multiprocess=0.70.16
murmurhash=1.0.10
narwhals=1.31.0
nbformat=5.10.4
ncurses=6.5
nest-asyncio=1.6.0
networkx=3.4.2
nlohmann_json=3.11.3
nomkl=1.0
numpy=2.2.4
openssl=3.4.1
optree=0.14.1
orc=2.1.1
packaging=24.2
pandas=2.2.3
parso=0.8.4
pexpect=4.9.0
pickleshare=0.7.5
pillow=11.1.0
pip=25.0.1
pkgutil-resolve-name=1.3.10
platformdirs=4.3.7
plotly=6.0.1
preshed=3.0.9
prometheus-cpp=1.3.0
prompt-toolkit=3.0.50
propcache=0.2.1
psutil=7.0.0
ptyprocess=0.7.0
pure_eval=0.2.3
pyarrow=19.0.1
pyarrow-core=19.0.1
pybind11=2.13.6
pybind11-global=2.13.6
pycparser=2.22
pydantic=2.10.6
pydantic-core=2.27.2
pygments=2.19.1
pysocks=1.7.1
python=3.12.9
python-dateutil=2.9.0.post0
python-fastjsonschema=2.21.1
python-tzdata=2025.1
python-xxhash=3.5.0
python_abi=3.12
pytorch=2.6.0
pytz=2024.1
pyyaml=6.0.2
pyzmq=26.3.0
re2=2024.07.02
readline=8.2
referencing=0.36.2
regex=2024.11.6
requests=2.32.3
rich=13.9.4
rpds-py=0.23.1
sacremoses=0.0.53
safetensors=0.5.3
scikit-learn=1.6.1
scipy=1.15.2
setuptools=75.8.2
shellingham=1.5.4
six=1.17.0
sleef=3.8
smart-open=7.1.0
smart_open=7.1.0
snappy=1.2.1
spacy=3.8.2
spacy-alignments=0.9.1
spacy-curated-transformers=0.3.0
spacy-legacy=3.0.12
spacy-loggers=1.0.5
spacy-transformers=1.1.9
srsly=2.5.1
stack_data=0.6.3
sympy=1.13.1
thinc=8.3.2
threadpoolctl=3.6.0
tk=8.6.13
tokenizers=0.21.1
torchaudio=2.6.0
torchvision=0.21.0
tornado=6.4.2
tqdm=4.67.1
traitlets=5.14.3
transformers=4.17.0
typer=0.15.2
typer-slim=0.15.2
typer-slim-standard=0.15.2
typing-extensions=4.12.2
typing_extensions=4.12.2
tzdata=2025a
ujson=5.10.0
urllib3=2.3.0
wasabi=1.1.3
wcwidth=0.2.13
weasel=0.4.1
wheel=0.45.1
wrapt=1.17.2
xxhash=0.8.3
xz=5.6.4
xz-gpl-tools=5.6.4
xz-tools=5.6.4
yaml=0.2.5